Move LinearMap to new map module
